Enable DataRow/DataRowView adapters in PowerShell Core
